requirements.txt
# Основные зависимости
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.0
openpyxl==3.1.5

# Для разработки
pytest==8.3.3
black==24.10.0